change dtype of labels

+++ src/dataloader.py
@@ -11,10 +11,9 @@ class DrugTargetData(Dataset):
 
         x = self.data
         y = self.labels
-        # TODO: fix the bug: map everything to torch.bfloat16. Take care to map categorical features separately from
-        #  numerical features.
+
         self.features = torch.tensor(x, dtype=torch.float32)
-        self.labels = torch.tensor(y, dtype=torch.int64)
+        self.labels = torch.tensor(y, dtype=torch.float32)
 
     def __getitem__(self, index):
         return self.features[index], self.labels[index]
